Move `cancel_build_workflows` function to the `workflow` module
