+++ src/bors/handlers/workflow.rs
@@ -7,8 +7,9 @@ use crate::bors::handlers::labels::handle_label_trigger;
 use crate::bors::merge_queue::AUTO_BRANCH_NAME;
 use crate::bors::merge_queue::MergeQueueSender;
 use crate::bors::{FailedWorkflowRun, RepositoryState, WorkflowRun};
-use crate::database::{BuildStatus, WorkflowModel, WorkflowStatus};
+use crate::database::{BuildModel, BuildStatus, WorkflowModel, WorkflowStatus};
 use crate::github::LabelTrigger;
+use crate::github::api::client::GithubRepositoryClient;
 use octocrab::models::CheckRunId;
 use octocrab::models::workflows::{Conclusion, Job, Status};
 use octocrab::params::checks::CheckRunConclusion;
@@ -642,4 +643,39 @@ auto_build_failed = ["+foo", "+bar", "-baz"]
             })
             .await;
     }
+}
+
+pub async fn cancel_build_workflows(
+    client: &GithubRepositoryClient,
+    db: &PgDbClient,
+    build: &BuildModel,
+    check_run_conclusion: CheckRunConclusion,
+) -> anyhow::Result<Vec<octocrab::models::RunId>> {
+    let pending_workflows = db.get_pending_workflows_for_build(build).await?;
+    let pending_workflows: Vec<octocrab::models::RunId> = pending_workflows
+        .into_iter()
+        .map(|id| octocrab::models::RunId(id.0))
+        .collect();
+
+    tracing::info!("Cancelling workflows {:?}", pending_workflows);
+    client.cancel_workflows(&pending_workflows).await?;
+
+    db.update_build_status(build, BuildStatus::Cancelled)
+        .await?;
+
+    if let Some(check_run_id) = build.check_run_id {
+        if let Err(error) = client
+            .update_check_run(
+                CheckRunId(check_run_id as u64),
+                CheckRunStatus::Completed,
+                Some(check_run_conclusion),
+                None,
+            )
+            .await
+        {
+            tracing::error!("Could not update check run {check_run_id}: {error:?}");
+        }
+    }
+
+    Ok(pending_workflows)
 }
